Rename raster and raster_sequence modules.
Now called spectrogram and spectrogram_sequence.
